Refactoring of interceptor stacks / jms facade, closing: JBMESSAGING-156, JBMESSAGING-27, JBMESSAGING-161, JBMESSAGING-83, JBMESSAGING-84

+++ src/main/org/jboss/jms/message/StreamMessageDelegate.java
@@ -22,6 +22,8 @@
 package org.jboss.jms.message;
 
 import javax.jms.JMSException;
+import javax.jms.MessageNotReadableException;
+import javax.jms.MessageNotWriteableException;
 import javax.jms.StreamMessage;
 
 /**
@@ -40,132 +42,190 @@ public class StreamMessageDelegate extends MessageDelegate implements StreamMess
 
    public boolean readBoolean() throws JMSException
    {
+      if (!bodyReadOnly)
+         throw new MessageNotReadableException("The message body is writeonly");
+
       return ((StreamMessage)message).readBoolean();
    }
 
    public byte readByte() throws JMSException
    {
+      if (!bodyReadOnly)
+         throw new MessageNotReadableException("The message body is writeonly");
+
       return ((StreamMessage)message).readByte();
    }
    
    public int readBytes(byte[] value) throws JMSException
    {
+      if (!bodyReadOnly)
+         throw new MessageNotReadableException("The message body is writeonly");
+
       return ((StreamMessage)message).readBytes(value);
    }
    
    public char readChar() throws JMSException
    {
+      if (!bodyReadOnly)
+         throw new MessageNotReadableException("The message body is writeonly");
+
       return ((StreamMessage)message).readChar();
    }
 
    public double readDouble() throws JMSException
    {
+      if (!bodyReadOnly)
+         throw new MessageNotReadableException("The message body is writeonly");
+
       return ((StreamMessage)message).readDouble();
    }
 
    public float readFloat() throws JMSException
    {
+      if (!bodyReadOnly)
+         throw new MessageNotReadableException("The message body is writeonly");
+
       return ((StreamMessage)message).readFloat();
    }
    
    public int readInt() throws JMSException
    {
+      if (!bodyReadOnly)
+         throw new MessageNotReadableException("The message body is writeonly");
+
       return ((StreamMessage)message).readInt();
    }
 
    public long readLong() throws JMSException
    {
+      if (!bodyReadOnly)
+         throw new MessageNotReadableException("The message body is writeonly");
+
       return ((StreamMessage)message).readLong();
    }
    
    public Object readObject() throws JMSException
    {
+      if (!bodyReadOnly)
+         throw new MessageNotReadableException("The message body is writeonly");
+
       return ((StreamMessage)message).readObject();
    }
    
    public short readShort() throws JMSException
    {
+      if (!bodyReadOnly)
+         throw new MessageNotReadableException("The message body is writeonly");
+
       return ((StreamMessage)message).readShort();
    }
    
    public String readString() throws JMSException
    {
+      if (!bodyReadOnly)
+         throw new MessageNotReadableException("The message body is writeonly");
+
       return ((StreamMessage)message).readString();
    }
 
    public void reset() throws JMSException
    {
+      bodyReadOnly = true;
       ((StreamMessage)message).reset();
    }
 
    public void writeBoolean(boolean value) throws JMSException
    {
+      if (bodyReadOnly)
+         throw new MessageNotWriteableException("The message body is readonly");
       bodyChange();
       ((StreamMessage)message).writeBoolean(value);
    }
 
    public void writeByte(byte value) throws JMSException
    {
+      if (bodyReadOnly)
+         throw new MessageNotWriteableException("The message body is readonly");
       bodyChange();
       ((StreamMessage)message).writeByte(value);
    }
    
    public void writeBytes(byte[] value) throws JMSException
    {
+      if (bodyReadOnly)
+         throw new MessageNotWriteableException("The message body is readonly");
       bodyChange();
       ((StreamMessage)message).writeBytes(value);
    }
    
    public void writeBytes(byte[] value, int offset, int length) throws JMSException
    {
+      if (bodyReadOnly)
+         throw new MessageNotWriteableException("The message body is readonly");
       bodyChange();
       ((StreamMessage)message).writeBytes(value, offset, length);
    }
    
    public void writeChar(char value) throws JMSException
    {
+      if (bodyReadOnly)
+         throw new MessageNotWriteableException("The message body is readonly");
       bodyChange();
       ((StreamMessage)message).writeChar(value);
    }
    
    public void writeDouble(double value) throws JMSException
    {
+      if (bodyReadOnly)
+         throw new MessageNotWriteableException("The message body is readonly");
       bodyChange();
       ((StreamMessage)message).writeDouble(value);
    }
    
    public void writeFloat(float value) throws JMSException
    {
+      if (bodyReadOnly)
+         throw new MessageNotWriteableException("The message body is readonly");
       bodyChange();
       ((StreamMessage)message).writeFloat(value);
    }
    
    public void writeInt(int value) throws JMSException
    {
+      if (bodyReadOnly)
+         throw new MessageNotWriteableException("The message body is readonly");
       bodyChange();
       ((StreamMessage)message).writeInt(value);
    }
    
    public void writeLong(long value) throws JMSException
    {
+      if (bodyReadOnly)
+         throw new MessageNotWriteableException("The message body is readonly");
       bodyChange();
       ((StreamMessage)message).writeLong(value);
    }
    
    public void writeObject(Object value) throws JMSException
    {
+      if (bodyReadOnly)
+         throw new MessageNotWriteableException("The message body is readonly");
       bodyChange();
       ((StreamMessage)message).writeObject(value);
    }
    
    public void writeShort(short value) throws JMSException
    {
+      if (bodyReadOnly)
+         throw new MessageNotWriteableException("The message body is readonly");
       bodyChange();
       ((StreamMessage)message).writeShort(value);
    }
    
    public void writeString(String value) throws JMSException
    {
+      if (bodyReadOnly)
+         throw new MessageNotWriteableException("The message body is readonly");
       bodyChange();
       ((StreamMessage)message).writeString(value);
    }
